added menu and difficulty
# Conflicts:
#	gui.py
#	main.py
